tests/requirements.txt
wheel
setuptools
pkginfo
cryptography
base58
python-dateutil==2.8.2
pymavryk==3.12.2
GitPython
pytest-xdist
ragger[tests,speculos,ledgerwallet]==1.24.0
